+++ MathmagicianTest/NaturalNumbersTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Mathmagician;
+using System.Numerics;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace MathmagicianTest
@@ -26,8 +27,8 @@ namespace MathmagicianTest
         public void ShouldPrintFirstNumber()
         {
             NaturalNumbers nats = new NaturalNumbers();
-            int expected = 1;
-            int actual = nats.GetFirst();
+            BigInteger expected = 1;
+            BigInteger actual = nats.GetFirst();
             Assert.AreEqual(expected, actual);
         }
 
@@ -35,9 +36,9 @@ namespace MathmagicianTest
         public void ShouldPrintNextNumber()
         {
             NaturalNumbers nats = new NaturalNumbers();
-            int input = nats.GetFirst();
-            int expected = 2;
-            int actual = nats.GetNext(input);
+            BigInteger input = nats.GetFirst();
+            BigInteger expected = 2;
+            BigInteger actual = nats.GetNext(input);
             Assert.AreEqual(expected, actual);
         }
 
@@ -45,9 +46,9 @@ namespace MathmagicianTest
         public void ShouldPrintNextNumberInput3()
         {
             NaturalNumbers nats = new NaturalNumbers();
-            int input = 3;
-            int expected = 4;
-            int actual = nats.GetNext(input);
+            BigInteger input = 3;
+            BigInteger expected = 4;
+            BigInteger actual = nats.GetNext(input);
             Assert.AreEqual(expected, actual);
         }
 
@@ -55,8 +56,8 @@ namespace MathmagicianTest
         public void ShouldPrintSequenceOfNumbers()
         {
             NaturalNumbers nats = new NaturalNumbers();
-            int[] expected = new int[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
-            int[] actual = nats.GetSequence();
+            BigInteger[] expected = new BigInteger[] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+            BigInteger[] actual = nats.GetSequence();
             CollectionAssert.AreEqual(expected, actual);
         }
 
@@ -64,9 +65,9 @@ namespace MathmagicianTest
         public void ShouldPrintVariableSequenceOfNumbers()
         {
             NaturalNumbers nats = new NaturalNumbers();
-            int input = 8;
-            int[] expected = new int[] { 1, 2, 3, 4, 5, 6, 7, 8 };
-            int[] actual = nats.GetSequence(input);
+            BigInteger input = 8;
+            BigInteger[] expected = new BigInteger[] { 1, 2, 3, 4, 5, 6, 7, 8 };
+            BigInteger[] actual = nats.GetSequence(input);
             CollectionAssert.AreEqual(expected, actual);
         }
 
@@ -74,9 +75,9 @@ namespace MathmagicianTest
         public void ShouldProvideSameSequences()
         {
             NaturalNumbers nats = new NaturalNumbers();
-            int input = 10;
-            int[] sequencenolength = nats.GetSequence();
-            int[] sequencewithlength = nats.GetSequence(input);
+            BigInteger input = 10;
+            BigInteger[] sequencenolength = nats.GetSequence();
+            BigInteger[] sequencewithlength = nats.GetSequence(input);
             CollectionAssert.AreEqual(sequencenolength, sequencewithlength);
         }
 
@@ -84,7 +85,7 @@ namespace MathmagicianTest
         public void PrintSequenceOfNumbers()
         {
             NaturalNumbers nats = new NaturalNumbers();
-            int length = 7;
+            BigInteger length = 7;
             string expected = "1 2 3 4 5 6 7";
             string actual = nats.ToString(nats.GetSequence(length));
             Assert.AreEqual(expected, actual);
